+++ CHANGES.rst
@@ -45,6 +45,10 @@ Changelog
 -   Fix collation of INTEGRAL notices by getting the GCN ID from the IVORN
     field.
 
+-   Add criteria when choosing new preferred external event that real events
+    are preferred over those likely not to be real, and SNEWS events won't be
+    overwritten by GRBs.
+
 -   Update bilby to 2.0.0 and bilby_pipe to 1.0.8. Use ROQ bases dedicated for
     BNS mass region for low-mass signal. Switch to acceptance-walk method for
     dynesty sampling, and use naccept=10, nlive=500, and nparallel=2 for